requirement.txt
absl-py==1.3.0
aio-pika==8.3.0
aiogram==2.23.1
aiohttp==3.8.4
aiormq==6.6.4
aiosignal==1.3.1
alabaster==0.7.12
applaunchservices==0.2.1
appnope==0.1.3
APScheduler==3.9.1.post1
argh==0.26.2
asgiref==3.6.0
astropy==4.0.2
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==22.2.0
Babel==2.9.1
backcall==0.2.0
backports.shutil-get-terminal-size==1.0.0
backports.zoneinfo==0.2.1
bidict==0.22.1
bkcharts==0.2
boto==2.49.0
boto3==1.26.74
botocore==1.29.74
Bottleneck==1.3.2
brotlipy==0.7.0
CacheControl==0.12.11
cached-property==1.5.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
clyent==1.2.2
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confluent-kafka==1.9.2
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
cycler==0.11.0
cytoolz==0.11.0
dask==2022.10.2
decorator==5.1.1
defusedxml==0.6.0
Django==4.1.7
django-extensions==3.2.1
django-private-storage==3.0
djangorestframework==3.14.0
dnspython==1.16.0
docopt==0.6.2
docutils==0.16
elasticsearch==7.13.0
entrypoints==0.3
fastcache==1.1.0
fbmessenger==6.0.0
fire==0.5.0
Flask==1.1.2
flatbuffers==23.1.21
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gast==0.5.3
glob2==0.7
gmpy2==2.0.8
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.1
h5py==3.8.0
HeapDict==1.0.1
httptools==0.5.0
humanfriendly==10.0
idna==3.4
imagesize==1.2.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
inflection==0.5.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.5
jsonpickle==2.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-core==4.6.3
jupyterlab==2.2.6
jwt==1.3.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy-object-proxy==1.4.3
libarchive-c==2.9
libclang==15.0.6.1
llvmlite==0.34.0
locket==1.0.0
magic-filter==1.0.9
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
mattermostwrapper==2.2
mccabe==0.6.1
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.2
mpmath==1.1.0
msgpack==1.0.4
multidict==5.2.0
multipledispatch==0.6.0
names==0.3.0
networkx==2.6.3
numexpr==2.7.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pamqp==3.2.1
parso==0.7.0
partd==1.3.0
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
Pillow==9.4.0
pkginfo==1.6.1
pkgutil-resolve-name==1.3.10
pluggy==1.0.0
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.28
protobuf==3.19.6
psycopg2-binary==2.9.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.21
pycurl==7.43.0.6
pydantic==1.10.2
pydot==1.4.2
pyflakes==2.2.0
PyJWT==2.6.0
pykwalify==1.8.0
pymongo==3.10.1
pyodbc===4.0.0-unsupported
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==0.0.0
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==19.0.2
QDarkStyle==2.8.1
QtPy==1.9.0
questionary==1.10.0
randomname==0.1.5
randomtimestamp==2.0.0
rasa==3.4.4
rasa-sdk==3.4.0
redis==4.5.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rocketchat-API==1.28.1
rsa==4.9
Rtree==0.9.4
ruamel-yaml==0.15.87
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-image==0.17.2
scikit-learn==1.1.3
scipy==1.8.1
Send2Trash==1.5.0
sentry-sdk==1.11.1
simplegeneric==0.8.1
six==1.16.0
sklearn-crfsuite==0.3.6
slack-sdk==3.20.0
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
soupsieve==2.0.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.4.46
sqlparse==0.4.3
tables==3.6.1
tabulate==0.9.0
tarsafe==0.0.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.4
tensorflow-addons==0.17.1
tensorflow-estimator==2.8.0
tensorflow-hub==0.12.0
tensorflow-text==2.8.2
termcolor==2.2.0
terminado==0.9.1
terminaltables==3.1.10
testpath==0.4.4
threadpoolctl==3.1.0
toolz==0.12.0
tornado==6.0.4
tqdm==4.64.1
twilio==7.14.2
typeguard==2.13.3
typing-extensions==4.5.0
typing-utils==0.1.0
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.14
uvloop==0.17.0
wcwidth==0.2.6
webencodings==0.5.1
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==2.2.3
widgetsnbextension==3.5.1
wrapt==1.14.1
wt==2.1.1
xlrd==1.2.0
xlwings==0.20.8
xlwt==1.3.0
yarl==1.8.2
zict==2.0.0
zipp==3.14.0
zope.event==4.5.0
